requirements.txt
3to2==1.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.0.2
Flask-Script==2.0.6
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
Pillow==5.3.0
PyAutoGUI==0.9.41
pydgutils==0.0.16
PyGetWindow==0.0.4
PyMsgBox==1.0.6
pypiwin32==219
PyRect==0.1.4
PyScaffold==2.5.11
PyScreeze==0.1.19
PyTweening==1.0.3
rabird.core==0.3.17
rabird.winio==0.2.6
requests==2.21.0
selenium==3.9.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.2
winiobinary==0.0.6
